requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
bcrypt>=4.0.0
python-multipart>=0.0.12
pydantic>=2.10.0
python-dotenv>=1.0.0
slowapi>=0.1.9
email-validator>=2.0.0
psycopg2-binary>=2.9.9
pymysql>=1.1.0
httpx>=0.27.0
openai>=1.0.0
pgvector>=0.3.0
PyJWT>=2.8.0
numpy>=1.26.0
